DreamTracerBackend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
celery
redis
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic-settings
google-generativeai
sentence-transformers
networkx
numpy
scipy
scikit-learn
pytest
pytest-asyncio
httpx